+++ packages/editor-common/web/src/Base/createBasePlugin.jsx
@@ -1,4 +1,4 @@
-import { includes, get } from 'lodash';
+import { includes } from 'lodash';
 import createBaseComponent from './createBaseComponent';
 import createToolbar from './createBaseToolbar';
 import createInsertPluginButton from './createBaseInsertPluginButton';
@@ -39,12 +39,11 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
   } = config;
   const toolbarTheme = { ...getToolbarTheme(config.theme, 'plugin'), ...config.theme };
   const Toolbar =
-    config.toolbar &&
-    config.toolbar.InlineButtons &&
+    config?.toolbar?.InlineButtons &&
     createToolbar({
       buttons: {
         all: config.toolbar.InlineButtons,
-        hidden: get(settings, 'toolbar.hidden', []),
+        hidden: settings?.toolbar?.hidden || [],
       },
       theme: { ...toolbarTheme, ...config.theme },
       pubsub,
@@ -54,16 +53,14 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
       anchorTarget,
       relValue,
       t,
-      name: config.toolbar.name,
+      name: config?.toolbar?.name,
       uiSettings: config.uiSettings,
       getToolbarSettings: config.getToolbarSettings,
       getEditorBounds,
     });
   const InsertPluginButtons =
     settings.showInsertButtons &&
-    config.toolbar &&
-    config.toolbar.InsertButtons &&
-    config.toolbar.InsertButtons.map(button => ({
+    config?.toolbar?.InsertButtons?.map(button => ({
       buttonSettings: button,
       component: createInsertPluginButton({
         blockType: config.type,