+++ packages/plugin-button/web/src/toolbar/inline-buttons.jsx
@@ -1,4 +1,3 @@
-import { get } from 'lodash';
 import {
   BUTTONS,
   getModalStyles,
@@ -22,7 +21,7 @@ const MobileFullScreenCustomStyle = {
 
 export default ({ settings, isMobile }) => {
   const customStyles = isMobile ? MobileFullScreenCustomStyle : DesktopCustomModalStyles;
-  const icon = get(settings, 'toolbar.icons.advanced_settings', PluginSettingsIcon);
+  const icon = settings?.toolbar?.icons?.['advanced_settings'] || PluginSettingsIcon;
   return [
     { keyName: 'sizeSmallLeft', type: BUTTONS.SIZE_SMALL_LEFT, mobile: false },
     { keyName: 'sizeSmallCenter', type: BUTTONS.SIZE_SMALL_CENTER, mobile: false },