+++ packages/plugin-code-block/web/src/createCodeBlockPlugin.js
@@ -1,5 +1,4 @@
 import CodeUtils from 'draft-js-code';
-import { get } from 'lodash';
 import { createBasePlugin } from 'wix-rich-content-editor-common';
 import { CODE_BLOCK_TYPE } from './types';
 // import PrismDecorator from './PrismDecorator';
@@ -56,7 +55,7 @@ const createUnderlyingPlugin = (/*{ theme }*/) => ({
 
 const createCodeBlockPlugin = (config = {}) => {
   const type = CODE_BLOCK_TYPE;
-  const icon = get(config, 'code-block.toolbar.icons.codeBlock');
+  const icon = config?.['code-block']?.toolbar?.icons?.codeBlock;
   const {
     helpers,
     theme,