+++ packages/plugin-file-upload/web/src/toolbar/insert-file-upload-buttons.jsx
@@ -1,10 +1,9 @@
-import { get } from 'lodash';
 import { TOOLBARS } from 'wix-rich-content-editor-common';
 import { DEFAULTS } from '../file-upload-component';
 import { InsertPluginIcon } from '../icons';
 
 export default ({ helpers, settings, t }) => {
-  const icon = get(settings, 'toolbar.icons.UploadFile', InsertPluginIcon);
+  const icon = settings?.toolbar?.icons?.UploadFile || InsertPluginIcon;
   return [
     {
       type: 'file',