+++ packages/plugin-map/web/src/toolbar/inline-buttons.jsx
@@ -1,4 +1,3 @@
-import { get } from 'lodash';
 import {
   BUTTONS,
   getModalStyles,
@@ -13,13 +12,13 @@ const getAlignmentButtonPropsFn = getEditorBounds => ({ componentData }) => {
   const editorBounds = getEditorBounds();
   const maxAlignmentWidth = editorBounds ? editorBounds.width - 1 : MAX_ALIGNMENT_WIDTH;
   return {
-    disabled: get(componentData, 'config.width', 0) > maxAlignmentWidth,
+    disabled: (componentData?.config?.width || 0) > maxAlignmentWidth,
   };
 };
 
 export default ({ settings, t, helpers, getEditorBounds, isMobile }) => {
   const { maxWidth, minWidth, maxHeight, minHeight } = settings;
-  const icons = get(settings, 'toolbar.icons', {});
+  const icons = settings?.toolbar?.icons || {};
   return [
     {
       type: BUTTONS.WIDTH,