+++ packages/plugin-sound-cloud/web/src/toolbar/insert-buttons.jsx
@@ -1,4 +1,3 @@
-import { get } from 'lodash';
 import { DEFAULTS } from '../soundCloud';
 import { getModalStyles, TOOLBARS } from 'wix-rich-content-editor-common';
 import SoundCloudURLInputModal from './soundCloudURLInputModal';
@@ -15,7 +14,7 @@ export default ({ helpers, t, isMobile, settings }) => {
   if (isMobile) {
     customStyles.content.minHeight = '202px';
   }
-  const icon = get(settings, 'toolbar.icons.SoundCloud', InsertPluginIcon);
+  const icon = settings?.toolbar?.icons?.SoundCloud || InsertPluginIcon;
   return [
     {
       type: 'modal',