+++ packages/plugin-text-color/web/src/toolbar/TextColorButton.jsx
@@ -4,7 +4,6 @@ import TextColorIcon from './TextColorIcon';
 import { TEXT_COLOR_TYPE } from '../types';
 import BaseTextColor from './BaseTextColor';
 import { textForegroundPredicate } from '../text-decorations-utils';
-import { get } from 'lodash';
 
 export default class TextColorButton extends Component {
   constructor(props) {
@@ -14,11 +13,11 @@ export default class TextColorButton extends Component {
 
   render() {
     const settings = this.props.config[TEXT_COLOR_TYPE];
-    const icon = get(settings, 'toolbar.icons.TextColor');
+    const iconTextColor = settings?.toolbar?.icons?.TextColor || TextColorIcon;
     const pluginParams = {
       dataHook: 'TextColorButton',
       toolTip: 'TextColorButton_Tooltip',
-      icon: icon || TextColorIcon,
+      icon: iconTextColor,
       type: TEXT_COLOR_TYPE,
       predicate: textForegroundPredicate,
     };